MFC r356138: Reimplement gvinum orphanization.
gvinum was the only GEOM class, using consumer nstart/nend fields. Making
it do its own accounting for orphanization purposes allows in perspective
to remove burden of that expensive for SMP accounting from GEOM.

Also the previous implementation spinned in a tight event loop, waiting
for all active BIOs to complete, while the new one knows exactly when it
is possible to close the consumer.


git-svn-id: https://svn.freebsd.org/base/stable/11@356577 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f