MFC r356279: libbe(3): promote dependent clones when destroying a BE
When removing a boot environment iterate over the dependents and process the
snapshots by grabbing any clones. Promote the clones we found and then
remove the target environment.

This fixes the ability to destroy a boot environment when it has been used
to spawn one or more other boot environments.

PR:		242592


git-svn-id: https://svn.freebsd.org/base/stable/11@356593 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f