MFC r356703
vmx: fix initialization of TSO related descriptor fields

Fix a mistake introduced by r343291, which ported the vmx(4)
driver to iflib.
In case of TSO, the hlen field of the (first) tx descriptor must
be initialized to the cumulative length of Ethernet, IP and TCP
headers. The length of the TCP header was missing.

PR:             236999
Reported by:    pkelsey
Reviewed by:    avg
Differential Revision:  https://reviews.freebsd.org/D22967


git-svn-id: https://svn.freebsd.org/base/stable/12@356932 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f