MFC r356474, r356480, r356482, r356506: Add Host Memory Buffer support to nvme(4).
This allows cheapest DRAM-less NVMe SSDs to use some of host RAM (about
1MB per 1GB on the devices I have) for its metadata cache, significantly
improving random I/O performance.  Device reports minimal and preferable
size of the buffer.  The code limits it to 5% of physical RAM by default.
If the buffer can not be allocated or below minimal size, the device will
just have to work without it.


git-svn-id: https://svn.freebsd.org/base/stable/12@356961 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f