The KASSERT is too strict: revert r357897
It's valid for a periph to be removed with outstanding transactions on the
device. In CAM, multiple periphs attach to a single device. There's no interlock
to prevent one of these going away while other periphs have outstanding CCBs and
it's not an error either. Remove this overly agressive KASSERT to prevent
false-positive panics when devices depart.


git-svn-id: https://svn.freebsd.org/base/head@357969 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f