MFC: r357149 Fix a crash in the NFSv4 server.
The PR reported a crash that occurred when a file was removed while
client(s) were actively doing lock operations on it.
Since nfsvno_getvp() will return NULL when the file does not exist,
the bug was obvious and easy to fix via this patch. It is a little
surprising that this wasn't found sooner, but I guess the above
case rarely occurs.

PR:		242768


git-svn-id: https://svn.freebsd.org/base/stable/12@358036 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f